Piano Code & UI Code
There were a few bugs I ran into when trying to use the piano code. Specifically within the UI API and the actual setup of the piano. I fixed the piano so that it is actually accurate and plays correct notes, but when you use require("ui") it will either break the order of the notes on the keyboard or not compile at all.